#35 dependencies upgraded  + small refactorings
