+++ CustomerIntelligence/Tests/CustomerIntelligence.Querying.Tests/packages.config
@@ -7,9 +7,9 @@
   <package id="Microsoft.AspNet.OData" version="5.8.0" targetFramework="net451" />
   <package id="Microsoft.AspNet.WebApi.Client" version="5.2.3" targetFramework="net451" />
   <package id="Microsoft.AspNet.WebApi.Core" version="5.2.3" targetFramework="net451" />
-  <package id="Microsoft.OData.Core" version="6.14.0" targetFramework="net451" />
-  <package id="Microsoft.OData.Edm" version="6.14.0" targetFramework="net451" />
-  <package id="Microsoft.Spatial" version="6.14.0" targetFramework="net451" />
+  <package id="Microsoft.OData.Core" version="6.15.0" targetFramework="net461" />
+  <package id="Microsoft.OData.Edm" version="6.15.0" targetFramework="net461" />
+  <package id="Microsoft.Spatial" version="6.15.0" targetFramework="net461" />
   <package id="Moq" version="4.2.1510.2205" targetFramework="net461" />
   <package id="Newtonsoft.Json" version="7.0.1" targetFramework="net461" />
   <package id="NMemory" version="1.0.1" targetFramework="net451" />
