+++ Querying/Querying.QueryExecution/packages.config
@@ -1,7 +1,7 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
   <package id="2GIS.NuClear.Metamodeling" version="2.0.0" targetFramework="net461" />
-  <package id="2GIS.NuClear.Utils" version="1.0.3" targetFramework="net461" />
+  <package id="2GIS.NuClear.Utils" version="1.0.12" targetFramework="net461" />
   <package id="EntityFramework" version="6.1.3" targetFramework="net451" />
-  <package id="Microsoft.OData.Edm" version="6.14.0" targetFramework="net451" />
+  <package id="Microsoft.OData.Edm" version="6.15.0" targetFramework="net461" />
 </packages>
