+++ Querying/Querying.Storage/packages.config
@@ -6,8 +6,8 @@
   <package id="2GIS.NuClear.Settings.API" version="1.2.0" targetFramework="net461" />
   <package id="2GIS.NuClear.Storage.API" version="2.0.0" targetFramework="net461" />
   <package id="2GIS.NuClear.Storage.EntityFramework" version="2.0.0" targetFramework="net461" />
-  <package id="2GIS.NuClear.Tracing.API" version="1.0.3" targetFramework="net461" />
-  <package id="2GIS.NuClear.Utils" version="1.0.3" targetFramework="net461" />
+  <package id="2GIS.NuClear.Tracing.API" version="1.0.4-logstash-bc8fb3-174" targetFramework="net461" />
+  <package id="2GIS.NuClear.Utils" version="1.0.12" targetFramework="net461" />
   <package id="EntityFramework" version="6.1.3" targetFramework="net461" />
   <package id="Microsoft.Data.Edm" version="5.7.0" targetFramework="net461" />
 </packages>
