+++ Querying/Querying.Web.OData/Web.config
@@ -9,7 +9,7 @@
     <add key="TargetEnvironmentName" value="Dev" />
   </appSettings>
   <connectionStrings>
-    <add name="CustomerIntelligence" providerName="System.Data.SqlClient" connectionString="Data Source=uk-sql01;Initial Catalog=CustomerIntelligence20;Integrated Security=True" />
+    <add name="CustomerIntelligence" providerName="System.Data.SqlClient" connectionString="Data Source=uk-sql01;Initial Catalog=CustomerIntelligence21;Integrated Security=True" />
     <add name="Logging" connectionString="http://uk-erm-kibana01:8194" />
   </connectionStrings>
   <system.web>
@@ -36,15 +36,15 @@
     <assemblyBinding xmlns="urn:schemas-microsoft-com:asm.v1">
       <dependentAssembly>
         <assemblyIdentity name="Microsoft.OData.Core" publicKeyToken="31bf3856ad364e35" culture="neutral" />
-        <bindingRedirect oldVersion="0.0.0.0-6.14.0.0" newVersion="6.14.0.0" />
+        <bindingRedirect oldVersion="0.0.0.0-6.15.0.0" newVersion="6.15.0.0" />
       </dependentAssembly>
       <dependentAssembly>
         <assemblyIdentity name="Microsoft.OData.Edm" publicKeyToken="31bf3856ad364e35" culture="neutral" />
-        <bindingRedirect oldVersion="0.0.0.0-6.14.0.0" newVersion="6.14.0.0" />
+        <bindingRedirect oldVersion="0.0.0.0-6.15.0.0" newVersion="6.15.0.0" />
       </dependentAssembly>
       <dependentAssembly>
         <assemblyIdentity name="Microsoft.Spatial" publicKeyToken="31bf3856ad364e35" culture="neutral" />
-        <bindingRedirect oldVersion="0.0.0.0-6.14.0.0" newVersion="6.14.0.0" />
+        <bindingRedirect oldVersion="0.0.0.0-6.15.0.0" newVersion="6.15.0.0" />
       </dependentAssembly>
       <dependentAssembly>
         <assemblyIdentity name="Microsoft.Owin" publicKeyToken="31bf3856ad364e35" culture="neutral" />
