+++ Querying/Querying.Web.OData/packages.config
@@ -27,14 +27,14 @@
   <package id="Microsoft.AspNet.WebApi.Cors" version="5.2.3" targetFramework="net451" />
   <package id="Microsoft.AspNet.WebApi.Owin" version="5.2.3" targetFramework="net451" />
   <package id="Microsoft.AspNet.WebApi.WebHost" version="5.2.3" targetFramework="net461" />
-  <package id="Microsoft.OData.Core" version="6.14.0" targetFramework="net461" />
-  <package id="Microsoft.OData.Edm" version="6.14.0" targetFramework="net461" />
+  <package id="Microsoft.OData.Core" version="6.15.0" targetFramework="net461" />
+  <package id="Microsoft.OData.Edm" version="6.15.0" targetFramework="net461" />
   <package id="Microsoft.Owin" version="3.0.1" targetFramework="net451" />
   <package id="Microsoft.Owin.Host.SystemWeb" version="3.0.1" targetFramework="net451" />
-  <package id="Microsoft.Spatial" version="6.14.0" targetFramework="net461" />
+  <package id="Microsoft.Spatial" version="6.15.0" targetFramework="net461" />
   <package id="Newtonsoft.Json" version="7.0.1" targetFramework="net461" />
   <package id="Owin" version="1.0" targetFramework="net451" />
-  <package id="Swashbuckle.Core" version="5.2.2" targetFramework="net451" />
-  <package id="Swashbuckle.OData" version="2.16.0" targetFramework="net461" />
+  <package id="Swashbuckle.Core" version="5.3.2" targetFramework="net461" />
+  <package id="Swashbuckle.OData" version="2.17.1" targetFramework="net461" />
   <package id="Unity" version="4.0.1" targetFramework="net461" />
 </packages>
